requirements.txt
django==2.2.24
djangorestframework==3.13.1
